MessageUpdater: I'm old!
