+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from "./intrinsics/index.js";
 export * from "./manager/index.js";
-export * from "./payload/index.js";
+export * from "./payload/PayloadBuilder.js";
 export * from "./reconciler/index.js";
 export * from "./renderer/index.js";
 export * from "./updater/index.js";
