--- src/payload/index.ts
@@ -1,2 +1,0 @@
-export * from './PayloadBuilder.js';
-export * from './types.js';
