+++ src/renderer/DJSXRenderer.ts
@@ -5,7 +5,7 @@ import type { DJSXRendererEventMap, DJSXRendererOptions } from "./types.js";
 import { type InternalNode, JSXRenderer } from "../reconciler/index.js";
 import type { DJSXEventHandlerMap } from "../intrinsics/index.js";
 import { MessageUpdater, REPLY_TIMEOUT, type MessageUpdateable } from "../updater/index.js";
-import { PayloadBuilder } from "../payload/index.js";
+import { PayloadBuilder } from "../payload/PayloadBuilder.js";
 import { resolveFile } from "../utils/resolve.js";
 import { defaultLog } from "../utils/log.js";
