+++ src/updater/index.ts
@@ -1,3 +1,3 @@
-export * from "./MessageUpdater.js"
+export * from "./MessageUpdaterOld.js"
 export * from "./types.js"
 export * from "./utils.js"
