+++ src/utils/log.ts
@@ -1,4 +1,6 @@
-export function defaultLog(level: 'message' | 'warn' | 'error' | 'trace', category: string, message: string, ...args: any[]) {
+export type ILogger = (level: 'message' | 'warn' | 'error' | 'trace', category: string, message: string, ...args: any[]) => void;
+
+export const defaultLog: ILogger = (level, category, message, ...args) => {
     switch (level) {
         case 'message':
             console.log(`[${category}] ${message}`, ...args);
